Add ruff (lint + format) and ty (type checking) to the project
Adds [tool.ruff] and [tool.ty] config to pyproject.toml with a curated
ruff rule set (E, F, W, I, UP, B, C4, SIM, RUF, ASYNC, LOG, G, PIE, RET,
PERF, N), pins ruff/ty in requirements-dev.txt, reformats the whole tree
with `ruff format`, and fixes the pre-existing lint and type-check debt
those tools surfaced so both run clean.

Production-code type fixes include: HA's ConfigFlowResult vs. the
generic FlowResult in config_flow.py, narrowing BoundEntity.desc to its
platform-specific subclass (SelectDesc/NumberDesc/SensorDesc/etc.) via
cast() instead of an unchecked annotation, converting HA device_class
strings to their proper enum types, a resolve_registry callback typed
as `object` instead of `DeviceRegistry | None`, and a couple of other
narrow correctness fixes (CA key type validation, an index-out-of-bounds
false positive from an empty-tuple fallback, a bool/dict argument swap).

Test-file fixes are mechanical: narrowing SamsungEntityDescription to
the correct subclass via isinstance()/cast() before accessing
subclass-only fields, and asserting Optional write_fn/unit_fn fields
are set before calling them.

+++ custom_components/localthings/registry/capabilities/dehumidifier.py
@@ -7,6 +7,7 @@ humidity -- not temperature -- is this device's primary control, and there
 is no climate composite: power, mode, and humidity are exposed as separate
 entities rather than folded into one card.
 """
+
 from ..capability import Capability
 from ..entities import NumberDesc, SelectDesc, SensorDesc
 from .common import int_or_none
@@ -16,21 +17,26 @@ def _first_mode(rep):
     """Representative scalar for the operating-mode select. `modes` is a
     single-element list on every dump seen so far, mirroring
     airconditioner._first_mode's handling of the same field shape."""
-    modes = rep.get('x.com.samsung.da.modes')
+    modes = rep.get("x.com.samsung.da.modes")
     if isinstance(modes, (list, tuple)):
         return modes[0] if modes else None
     return modes
 
 
 MODE = Capability(
-    href='/mode/vs/0',
-    poll_tier='warm',
+    href="/mode/vs/0",
+    poll_tier="warm",
     entities=(
-        SelectDesc(key='operating_mode', rep_fn=_first_mode,
-                   icon='mdi:tune-variant',
-                   options_field='x.com.samsung.da.supportedModes',
-                   write_fn=lambda p, rep, href=None: (
-                       ['mode', 'vs', '0'], {'x.com.samsung.da.modes': [p]})),
+        SelectDesc(
+            key="operating_mode",
+            rep_fn=_first_mode,
+            icon="mdi:tune-variant",
+            options_field="x.com.samsung.da.supportedModes",
+            write_fn=lambda p, rep, href=None: (
+                ["mode", "vs", "0"],
+                {"x.com.samsung.da.modes": [p]},
+            ),
+        ),
     ),
 )
 
@@ -43,21 +49,31 @@ MODE = Capability(
 # dump's values" section). Step comes live from the device's own `increment`
 # field.
 HUMIDITY = Capability(
-    href='/humidity/vs/0',
-    poll_tier='warm',
+    href="/humidity/vs/0",
+    poll_tier="warm",
     entities=(
-        SensorDesc(key='humidity', field='x.com.samsung.da.humidity',
-                   device_class='humidity', unit='%', state_class='measurement',
-                   value_fn=int_or_none),
-        NumberDesc(key='target_humidity', field='x.com.samsung.da.desiredHumidity',
-                   device_class='humidity', unit='%',
-                   icon='mdi:water-percent',
-                   entity_category='config',
-                   value_fn=int_or_none,
-                   step_fn=lambda rep: int_or_none(rep.get('increment')) or 1,
-                   write_fn=lambda p, rep, href=None: (
-                       ['humidity', 'vs', '0'],
-                       {'x.com.samsung.da.desiredHumidity': str(int(round(float(p))))})),
+        SensorDesc(
+            key="humidity",
+            field="x.com.samsung.da.humidity",
+            device_class="humidity",
+            unit="%",
+            state_class="measurement",
+            value_fn=int_or_none,
+        ),
+        NumberDesc(
+            key="target_humidity",
+            field="x.com.samsung.da.desiredHumidity",
+            device_class="humidity",
+            unit="%",
+            icon="mdi:water-percent",
+            entity_category="config",
+            value_fn=int_or_none,
+            step_fn=lambda rep: int_or_none(rep.get("increment")) or 1,
+            write_fn=lambda p, rep, href=None: (
+                ["humidity", "vs", "0"],
+                {"x.com.samsung.da.desiredHumidity": str(round(float(p)))},
+            ),
+        ),
     ),
 )
 
@@ -69,17 +85,17 @@ HUMIDITY = Capability(
 # families' schemas.
 # ---------------------------------------------------------------------------
 _DHM_IGNORED = [
-    '/availablecontrolsets/vs/0',  # opaque hex-encoded control-set bitmap (id: DHM)
-    '/da/softreset/vs/0',          # soft-reset trigger plumbing
-    '/keepnormalstate/vs/0',       # internal keep-normal flag
-    '/personality/presence/vs/0',  # presence-personalization plumbing (empty item value)
-    '/reserverulesets/vs/0',       # opaque hex-encoded schedule reservation blob
-    '/sensors/vs/0',               # empty {} on this dump
-    '/welcome/humidity/vs/0',      # welcome-mode plumbing (requestId/operatingStatus, inert)
+    "/availablecontrolsets/vs/0",  # opaque hex-encoded control-set bitmap (id: DHM)
+    "/da/softreset/vs/0",  # soft-reset trigger plumbing
+    "/keepnormalstate/vs/0",  # internal keep-normal flag
+    "/personality/presence/vs/0",  # presence-personalization plumbing (empty item value)
+    "/reserverulesets/vs/0",  # opaque hex-encoded schedule reservation blob
+    "/sensors/vs/0",  # empty {} on this dump
+    "/welcome/humidity/vs/0",  # welcome-mode plumbing (requestId/operatingStatus, inert)
     # Only supportedModes ([Off, Sleep]) is present -- no live "current
     # value" field on this dump to confirm the read/write contract, so per
     # the 'don't guess' rule this is left unmodeled rather than assumed.
-    '/mode/convenient/vs/0',
+    "/mode/convenient/vs/0",
 ]
 
 COVERAGE = [Capability(href=h) for h in _DHM_IGNORED]
